Keep apply_patch diff state in consumer
